Initial commit
Now that rdrview mostly seems to be working, development is likely to
become less erratic. So it's time to get organized and start keeping a
real git history. Just in case, I will preserve the original commits in
a local branch, but they are too messy to be made public.

Rdrview is mainly a transpilation done by hand of Mozilla's readability
library [1]. Development so far has been based on commit 52ab9b5c8916,
but the plan is to update it once in a while, to follow upstream and to
try to keep the output similar.

[1] https://github.com/mozilla/readability