Fix build for older versions of gcc
It's been reported that the build currently fails on gcc 5.4.0:

  https://github.com/eafer/rdrview/issues/1

The reason is that older versions of gcc don't allow constant pointers
to constant strings to be used for initialization. Use macros instead,
which are probably more portable in general.

The build with this patch has been tested successfully on gcc 4.9.2. It
should be noted that it throws a "missing braces around initializer"
warning. This appears to be a bug in the compiler:

  https://gcc.gnu.org/bugzilla/show_bug.cgi?id=53119