Make sure stdin is a tty before opening a browser
When rdrview spawns a web browser to render an article, the stdin file
is preserved. So, if the original HTML was read from a pipe, this same
pipe will be the stdin for the browser. This is not a problem for lynx,
because it actually supports some input via pipes. After handling such
input, it takes care of reopening stdin as the controlling terminal, so
that the TUI can work correctly.

But not all browsers appear to expect this. Links gets confused and
assumes that the pipe is the actual terminal input; anything left in the
pipe is understood as keystrokes by the TUI. Luckily the rdrview stdin
is empty by the time links is spawned, but the EOF still causes the
browser to close immediately, rendering nothing.

To fix this, reopen stdin as the controlling terminal ourselves, instead
of trusting that the browser will do it.