Correct base url option in man page examples
At one point during development, the command line flag for the base url
used to be "-b", but it was later changed to "-u" to avoid confusion
with the more important "-B" option. Most of the man page was updated
accordingly, but I forgot about the list of examples. Fix it.