Make sandbox code system-dependent
As a first step for supporting other *nix systems, move all seccomp code
to a new file and make it build conditionally, only on Linux. Also amend
the Makefile so that it only links to seccomp there.

On other systems, rdrview will just fail and inform the user that the
sandbox is not implemented, and that disabling it is an option (though
risky). Of course, the intention is to eventually provide a sandbox on
as many systems as possible.