Get rid of strcasestr()
It has been reported that rdrview can be made to build on OpenBSD, if it
drops my definition of strcasestr(), which clashes with the one from
their library:

  https://old.reddit.com/r/commandline/comments/jaluzg/firefox_reader_view_as_a_linux_command_line_tool/g9abaf9/

The irony is that strcasestr() was redefined because I was avoiding the
glibc version, which I thought would not be portable. Just call the one
provided by libxml2 directly. Add some documentation about the possible
encoding issues as well.