Link to iconv explicitly on FreeBSD
On FreeBSD, the linker is failing to find the functions from libiconv,
so modify the Makefile to link to that library explicitly.

I don't know much about FreeBSD, so this may not be optimal. According
to the documentation, the system has both a native version of iconv and
a "port", and I fear my Makefile may be using the port. It builds with
no issues though, so I'll ignore it for now.