Allow the "URL" to be a local path
For the command line interface of all browsers I've checked, the user is
allowed to supply a local filepath instead of a real URL. The way this
works is that the browser always checks first if a local file with the
given name exists.

Consistency here is important because browsers are regularly called by
other software, for instance via the BROWSER environment variable.
Rdrview is not exactly a browser but it will be used like one, so follow
the rest in this matter.

Since local files can now be processed in the natural way, without using
stdin, remove that example from the readme.