Include a reasonable user-agent string in requests
Shehu Dikko has reported via email that rdrview fails to process this
page:

  https://www.chronicle.com/article/why-do-we-thinking-learning-about-history-can-make-us-better

The problem is that the Incapsula CDN refuses to serve requests if their
user-agent string is not of a certain form. After some experimentation,
I believe what they want is something like

  <product>/<product-version> <product>/<product-version>

Use "Mozilla/5.0" for the first part, since that's the least conspicuous
choice. For the second part, be honest and use "rdrview/0.1".