Don't use static allocation for mailcap queries
The arrays used to store each line and field of a mailcap file are
currently allocated statically. There is no good reason for this:
mailcap parsing was a late addition and I was a bit tired of the project
by then, so I got lazy. Use the heap instead, to avoid wasting all that
memory.