Implement a sandbox for FreeBSD
Use the "capability mode" provided by Capsicum to sandbox rdrview on
FreeBSD. The sandboxed code is only allowed to work with existing open
file descriptors, namely the two temporary files for the process plus
standard I/O.

The amount of syscalls allowed by default is staggering, but I've
reviewed the list and they all appear to be harmless in this context.
Restricting them would just overcomplicate the code.

By the way, the nice() syscall is allowed by Capsicum, so the current
implementation of assert_sandbox_works() is no good. Replace nice() with
a call to fopen(), which should get blocked on FreeBSD as well and is a
better demonstration of what I'm trying to prevent. Make the expected
errno system-dependent.