Never delete the root node
Rdrview segfaults when it tries to parse any section of the ebook for
"The Rust Programming Language", for example

  https://doc.rust-lang.org/book/title-page.html

The irony does not escape me, but this is just a NULL pointer
dereference, so it wouldn't be a big deal even without a sandbox.

The problem with these pages is that the root node has a class which
includes the word "sidebar", so it's considered unlikely to be readable
and removed entirely. The existence of the root node is only verified
once, as soon as parse() is called, and the rest of the code trusts that
it continues to exist. This is reasonable, so prevent any deletion by
simply skipping the root node in all tree traversals.