Avoid NULL dereference on childless root nodes
Commit fed4ab48ac79 banned the root node from ever being deleted, since
that made no sense for the purpose of extracting content. A side effect
of this change was avoiding a NULL pointer dereference, but it's not
really fixed and it can still be triggered by a childless root node.

The problem here is that run_on_nodes() will always return NULL if it
hasn't iterated over any nodes, even when just a single iteration would
return a valid pointer. So don't assume that tops isn't NULL in
find_top_candidate().

Add a test for the bug as well.