Flush the file stream after downloading a webpage
Curl works with FILE pointers but our html parsing is done on mmapped
file descriptors, so it's important that none of the file's contents are
forgotten in the userland buffers. Call fflush() after the download is
done.

Without this patch, the extracted content from some pages will get
cropped. For example:

  https://doc.rust-lang.org/book/ch03-05-control-flow.html

It's a bit surprising that I took so long to stumble into this bug. I
think the reason is that a page's content is usually farther away from
the bottom of the document, so the crop would not affect it.