Run only sandboxed code inside the subprocess
The code that downloads the page would be tricky to sandbox, because it
requires access to a number of configuration files. I decided not to
attempt this: everything we do with libcurl is very ordinary, so for us
to be vulnerable there would have to be a critical bug in the library.
In that situation, attackers would have an endless number of far more
interesting targets than rdrview.

Nevertheless, the download is happening inside the subprocess which is
meant to be sandboxed, and inside a function called run_dangerous().
This is very confusing and, more importantly, it makes it impossible to
consider performing a second download after finding an HTML redirect.

So move the download to the parent process, and do the same with the
initialization of iconv, which has no need for a sandbox.