Move subprocess handling to a separate function
The code that creates a subprocess and obtains its exit status is
currently part of main(). I intend to run it in a loop along with
save_input_to_file() to handle html redirects, so move it to a separate
function to avoid excessive indentation.

It mostly decoupled well, the only problem was that valgrind complained
about 'command' getting leaked. It would be silly to pass it to the
child just so it can be freed, so make it static instead.