Don't keep open file streams during the fork
The child process and the parent both work on the same files, so the
userland file streams of the parent will often be invalid by the time
the child returns. This isn't a problem right now because the parent
doesn't work with these streams again, but that will change once html
redirects are implemented.

So close the streams before the fork and pass only descriptors to the
child.