Follow HTML redirects
I've received a report that rdrview fails to follow some redirects:

  https://github.com/eafer/rdrview/issues/11

The problem is caused by HTML redirects, which are naturally ignored
by libcurl. To handle them, make the child check for them early when
parsing the document; if they are found, save the target url to the
output file instead of the extracted text. Then make the parent retry
the download and fork a new child to parse it.