Don't use a file stream for input inside the fork
The parent process requires the input file to be opened as a stream so
that it can be passed to libcurl. But the child process only reads it
via mmap(), so it just needs a file descriptor. Mixing descriptors and
streams often brings problems, so don't reopen the stream after the
fork.