Reset the output file offset before reuse
Since commit 61756b1145e3, the output file may first store the
destination of HTML redirects, before getting reused for the actual
content. Each reuse requires resetting the file offset, but this is
currently being done to the input file instead. This error leaves a
number of null bytes at the beginning of the output, which at least
links will render into weird dots.

The mistake appears to have been caused by copy-paste of equivalent code
for the input file, so instead of fixing it in place, move both cases to
a new function.