Map the input file before setting up the sandbox
It's been reported that some standard libraries implement fstat() using
the newfstatat() syscall:

  https://github.com/eafer/rdrview/issues/14#issuecomment-775574784

This is a very annoying: if I allow newfstatat(), an attacker running
code inside the sandbox would be able to get information about any
filepath. Instead, restructure the code so that the only call to fstat()
happens before the sandbox is set up, and ban all variants of this
syscall (as well as mmap(), which is no longer needed either).