Escape base urls with invalid unicode
A user reports that links and images don't seems to work in non-english
pages:

  https://github.com/eafer/rdrview/issues/23

They actually work fine is the url is provided in the correct form, that
is, with percent-encoded unicode characters. But mainstream browsers do
accept invalid unicode urls, so users reasonably expect them to work in
rdrview as well.

There is no way to get libxml2 to be more permissive here, so instead
percent-encode the unicode characters before the url is needed. Since we
are working with the whole url here, be careful and don't mess with the
reserved characters.