Don't allow argc == 0
I just noticed that technically it would be possible for progname to get
set to NULL if argc is zero, at least for old versions of the Linux
kernel. I don't think this matters at all here, but it's always ugly to
throw a segfault when I can just exit. So, add a check before parsing
the arguments.