Update test for the Windows-1252 encoding
My last attempt at running tests run into problems with cp-1252. The
format of the text dump appears to have changed very slightly: it's now
missing a redundant separator row. This may have been caused by a new
version of links, or even by changes in libxml2, I guess. I haven't
checked, but it doesn't matter much: the only point of this test was to
verify that the encoding was understood, so the important part are the
left/right double quotation marks.

Just update the test with the new rendered result.