Allow futex() syscall in sandbox
The rdrview sandbox has been broken again for more than a year:

  https://github.com/eafer/rdrview/issues/15#issuecomment-1445158372

I can reproduce this myself since my last Ubuntu upgrade. The culprit
this time is the futex() syscall, which should have been obvious given
the error message. It seems harmless, so add it to the whitelist.

Take this chance to also allow fstat(). It's not really needed, but it
does get called for some reason, so allowing it makes the strace output
cleaner.