Don't call xmlMemSetup() on newer macos releases
Rdrview is not working on macos:

  https://github.com/eafer/rdrview/issues/47

The libxml2 maintainer pointed me to the macos release notes for 15.4,
which claim that the libxml2 memory allocation api is deprecated on that
platform:

  https://developer.apple.com/documentation/macos-release-notes/macos-15_4-release-notes

So introduce version checks and use malloc/free from the standard
library. Hopefully this won't break anything for homebrew libxml2 but I
don't know if anybody is using that.